feat: CSRF 설정 추가 ㅡ,ㅡ

+++ src/main/java/com/juyb99/dinorunboot/dto/request/ScoreRequestDTO.java
@@ -1,4 +1,4 @@
 package com.juyb99.dinorunboot.dto.request;
 
-public record ScoreRequestDTO(String nickname, int point) {
+public record ScoreRequestDTO(String nickname, Long point) {
 }
